services/image-moderation-api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
Pillow==10.1.0
numpy==1.24.3
torch==2.1.1
torchvision==0.16.1
transformers==4.35.2 
git+https://github.com/openai/CLIP.git#egg=clip
